Move `IStaticMod` to IModInterface.cs
