+++ Source/ac-library-csharp/Math/ModInt/IModInterface.cs
@@ -1,0 +1,66 @@
+﻿#if GENERIC_MATH
+using System.Numerics;
+#endif
+
+namespace AtCoder
+{
+    public interface IModInt<T>
+#if GENERIC_MATH
+     : INumberBase<T> where T : INumberBase<T>
+#endif
+    {
+        /// <summary>
+        /// 自身を x として、 xy≡1 なる y を返します。
+        /// </summary>
+        /// <remarks>
+        /// <para>制約: gcd(x, mod) = 1</para>
+        /// </remarks>
+        T Inv();
+
+        /// <summary>
+        /// 自身を x として、x^<paramref name="n"/> を返します。
+        /// </summary>
+        /// <remarks>
+        /// <para>制約: 0≤|<paramref name="n"/>|</para>
+        /// <para>計算量: O(log(<paramref name="n"/>))</para>
+        /// </remarks>
+        T Pow(ulong n);
+    }
+
+    /// <summary>
+    /// コンパイル時に決定する mod を表します。
+    /// </summary>
+    /// <example>
+    /// <code>
+    /// public readonly struct Mod1000000009 : IStaticMod
+    /// {
+    ///     public uint Mod => 1000000009;
+    ///     public bool IsPrime => true;
+    /// }
+    /// </code>
+    /// </example>
+    [IsOperator]
+    public interface IStaticMod
+    {
+        /// <summary>
+        /// mod を取得します。
+        /// </summary>
+        uint Mod { get; }
+
+        /// <summary>
+        /// mod が素数であるか識別します。
+        /// </summary>
+        bool IsPrime { get; }
+    }
+    public readonly struct Mod1000000007 : IStaticMod
+    {
+        public uint Mod => 1000000007;
+        public bool IsPrime => true;
+    }
+
+    public readonly struct Mod998244353 : IStaticMod
+    {
+        public uint Mod => 998244353;
+        public bool IsPrime => true;
+    }
+}
